requirements.txt
click==7.1.2
cmake==3.18.4.post1
dlib==19.21.0
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
face-recognition==1.3.0
face-recognition-models==0.3.0